commit task 100

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-long long int number = 612852475143;
+unsigned long int number = 612852475143;
 int oddPrimeNum = 3;
 int maxPrimeNum = 0;
 while (number % 2 == 0)
